リザルトオブジェクトがアクティブ時にリザルト遷移音を流す処理を修正
