--- Assets/Prefabs/SoundCanvas.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 912e461e014e5124aa864ecb0dda7223
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
