--- Assets/Script/OnResultObjectActiveCheck.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7f4736446edac1f47b728c7bf5e3b8c3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
